Don't pollute the source directory with binaries
tz's makefile can only build in-tree.  It *installs* to a different
directory (specified by `DESTDIR`), but the build artifacts are left in
the source tree, including `zic.o`, `zic`, etc.  This is disruptive to
CMake builds which are usually out-of-tree.  To work around this, copy
the whole source tree to a path under `CMAKE_CURRENT_BINARY_DIR`, and
build it from there.  It's a tiny project so the cost of copying is not
significant.